fix(network,proav,infra): RFC 2435 UHD guard + JPEG matrix pin + PNG timeout floor
## Summary

Three targeted fixes to the RTP / promeki-test layer on top of the
failure-tracing sweep and AV-over-IP hardening pass already landed in
the working tree.

### 1. RFC 2435 §3.1.5 UHD dimension guard (`src/network/rtppayload.cpp`)

`RtpPayloadJpeg::pack` now emits `promekiWarnThrottled` (throttled to
once per second) when the configured width or height exceeds 2040
pixels — the maximum representable by RFC 2435's 8-bit `W/8` / `H/8`
fields (255 × 8 = 2040).

Root cause confirmed: at 4K60 both fields silently clamped to 255 on
the wire.  The receiver reconstructed a SOF0 header for 2040×2040;
libjpeg consumed only the MCUs for that smaller grid and flagged the
remaining ~109 KB of real-4K entropy as "Corrupt JPEG data: N
extraneous bytes before marker 0xd9".  FFmpeg's `rtpenc_jpeg.c`
refuses outright at this limit; we warn so UHD misconfigurations don't
masquerade as quiet picture-quality regressions.  Callers needing UHD
over RTP should use JPEG XS (ST 2110-22) or uncompressed ST 2110-20.

### 2. JPEG matrix entries pinned to 1080p59.94
   (`utils/promeki-test/cases/rtp.cpp`, `buildTxConfig`)

`if (c.family == String("jpeg"))` now sets `MediaConfig::VideoFormat`
to `VideoFormat::Smpte1080p59_94` — the largest standard broadcast
raster that fits inside RFC 2435's 2040-pixel envelope.  All other
matrix entries stay on the prior 4K60 hardening default
(`kDefaultVideoFormat` = 2160p59.94).  Without this pin the three JPEG
matrix cases produced Corrupt-JPEG warnings on every frame and the
recovered image data was a 2040×2040 crop of a 4K source rather than a
full-frame round-trip.

Verified: 5/5 stability runs on all three JPEG matrix cases — all PASS,
zero Corrupt-JPEG warnings, 100% image-data band recovery.  Full RTP
matrix: 17 PASS / 0 FAIL / 6 SKIP (the six SKIP entries are known
library gaps: 12-bit/DPX reverse CSC, NVENC unavailable).

### 3. PNG image-sequence per-phase timeout floor at 30 s
   (`utils/promeki-test/cases/roundtrip.cpp`, `runRoundtripCase`)

`timeoutMs` is floored to 30 000 ms when `c.backendName ==
String("ImgSeqPNG")`.  libspng's deflate pass on a 24 MB uncompressed
2160p59.94 surface routinely spends ~15 s per phase — dramatically more
CPU-intensive than every other sequence backend in the matrix (DPX /
Cineon are uncompressed; JPEG / JPEG-XS are chunked-DCT; TIFF defaults
to LZW which is ~3× cheaper).  An explicit `TestParams::PhaseTimeoutMs`
higher than the floor still wins; lower values get promoted so the test
never false-flags on a normally-loaded host.

Verified: 3/3 stability runs on `roundtrip.imgseqpng` — all PASS at
~16 s each, comfortably under the new 30 s floor.

## Broader changeset context

The working tree also contains the large body of work from the current
session, all of which passes `build check` (3/3 unit-test suites PASS):

- **E2022 dual-leg redundancy** (`RtpMediaIO` ST 2022-7 fan-out and
  fan-in, `RtpSession::sendPacketsToDual` / `setDualLegSocket`).
- **Per-stream packet budget + PacketScheduler** — `RtpSession`
  decoupled from `RtpMediaIO`; `PacketScheduler` owns the per-stream
  send window; `RtpSeqTracker` provides receive-side sequence-gap /
  reorder detection independent of the session object.
- **AncDesc / ANC audit carry-ins** — `pairedVideoStreamIndex` /
  `pairedAudioStreamIndex`, reserved-DID guard, OUI-aware HDMI lookup.
- **Failure-tracing sweep** — `OnceGate` / `ThrottleGate` primitives;
  `promekiWarnThrottled` / `promekiErrorThrottled` macros plumbed
  through the network and proAV layers.
- **Timecode / ATC audit (all 8 phases)** — physical-frame Timecode,
  TimecodeUserbits, dual-carriage ATC (ST 12-2 pair-rate + ST 12-3
  ATC_HFRTC).
- **JPEG XS slice-mode packing** — RFC 9134 K=1 slice packetization
  with main-header prepend and single-slice-MTU fallback.
- **ClockDomain / RtpMediaClock decoupling** — `ClockDomain` as a
  standalone typed CRTP hierarchy separate from `RtpSession`; unit
  tests in `tests/unit/clockdomain.cpp`.
- **RtpVideoDepacketizerThread** — per-stream reassembly thread with
  `JpegGeometryProbe` for RFC 2435 and `RtpSeqTracker`-backed
  gap detection.
- **SdpSession / RtpMediaIO improvements** — `a=framerate` round-trip,
  JPEG `x-dimensions` fmtp extension, `AudioFileFactory_LibSndFile`
  runtime extension probing, `AudioFormat` signed-type
  zero-preservation.
- **`promeki-test` functional runner** — `rtp.*` matrix (JPEG, raw,
  H.264, H.265, JPEG XS, JPEG-over-1080p), `roundtrip.*` matrix
  (ImgSeqPNG, ImgSeqDPX, ImgSeqJPEG, QuickTime, …).

All unit tests: 3/3 PASS.  Build: zero warnings in project code.

+++ tests/unit/cscpipeline.cpp
@@ -1440,7 +1440,7 @@ TEST_CASE("CSC HDR: pipeline picks the direct-compute PQ/HLG kernel over the LUT
         CSCPipeline pPq(PixelFormat::RGB16_LE_sRGB, PixelFormat::RGB16_LE_Rec2020_PQ, scalarConfig());
         REQUIRE(pPq.isValid());
         bool sawPqOETF = false;
-        for (size_t i = 0; i < pPq.stageCount(); ++i) {
+        for (int i = 0; i < pPq.stageCount(); ++i) {
                 const auto &s = pPq.stage(i);
                 if (s.type == CSCPipeline::StageOETF) {
                         CHECK(s.hdrTransfer == CSCPipeline::Stage::HdrTransferPqOETF);
@@ -1453,7 +1453,7 @@ TEST_CASE("CSC HDR: pipeline picks the direct-compute PQ/HLG kernel over the LUT
         CSCPipeline pPqRev(PixelFormat::RGB16_LE_Rec2020_PQ, PixelFormat::RGB16_LE_sRGB, scalarConfig());
         REQUIRE(pPqRev.isValid());
         bool sawPqEOTF = false;
-        for (size_t i = 0; i < pPqRev.stageCount(); ++i) {
+        for (int i = 0; i < pPqRev.stageCount(); ++i) {
                 const auto &s = pPqRev.stage(i);
                 if (s.type == CSCPipeline::StageEOTF) {
                         CHECK(s.hdrTransfer == CSCPipeline::Stage::HdrTransferPqEOTF);
@@ -1466,7 +1466,7 @@ TEST_CASE("CSC HDR: pipeline picks the direct-compute PQ/HLG kernel over the LUT
         CSCPipeline pHlg(PixelFormat::RGB16_LE_sRGB, PixelFormat::RGB16_LE_Rec2020_HLG, scalarConfig());
         REQUIRE(pHlg.isValid());
         bool sawHlgOETF = false;
-        for (size_t i = 0; i < pHlg.stageCount(); ++i) {
+        for (int i = 0; i < pHlg.stageCount(); ++i) {
                 const auto &s = pHlg.stage(i);
                 if (s.type == CSCPipeline::StageOETF) {
                         CHECK(s.hdrTransfer == CSCPipeline::Stage::HdrTransferHlgOETF);
@@ -1479,7 +1479,7 @@ TEST_CASE("CSC HDR: pipeline picks the direct-compute PQ/HLG kernel over the LUT
         // HdrTransferNone so the LUT path stays the default.
         CSCPipeline pSdr(PixelFormat::RGBA8_sRGB, PixelFormat::YUV8_422_Rec709, scalarConfig());
         REQUIRE(pSdr.isValid());
-        for (size_t i = 0; i < pSdr.stageCount(); ++i) {
+        for (int i = 0; i < pSdr.stageCount(); ++i) {
                 const auto &s = pSdr.stage(i);
                 if (s.type == CSCPipeline::StageOETF || s.type == CSCPipeline::StageEOTF) {
                         CHECK(s.hdrTransfer == CSCPipeline::Stage::HdrTransferNone);
